requirements.txt
Flask>=2.2.2
python-dotenv
requests
Jinja2>=3.1.2